ref: move calc_test_id to shared
