+++ apps/worker/services/test_analytics/ta_timeseries.py
@@ -8,12 +8,12 @@ import test_results_parser
 from django.db import connections
 from django.db.models import Q
 
-from services.test_analytics.utils import calc_test_id
 from services.test_results import FlakeInfo
 from shared.django_apps.ta_timeseries.models import (
     Testrun,
     TestrunBranchSummary,
     TestrunSummary,
+    calc_test_id,
 )
 from shared.django_apps.test_analytics.models import Flake
